Added Scratch-compatible SB3, bug fixes, & more
I added a Scratch-compatible version of the engine and fixed a bug where it'll add you as the the wrong player. I also made it easier to use. It is now properly drag-and-drop use. Just make sure to click the green flag a couple times to set the variables and lists correctly.